% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classificationplots.R
\name{classificationplots}
\alias{classificationplots}
\title{Generate plots for classification}
\usage{
classificationplots(
  response,
  outcome,
  levels,
  cutoffs.1 = NULL,
  cutoffs.2 = NULL,
  cutoffs.3 = NULL,
  xlab.3 = "Predicted probability",
  labels.3 = NULL,
  xlim.3 = NULL,
  ylim.3 = c(0, 10)
)
}
\arguments{
\item{response}{continuous variable that is used as a classifier}

\item{outcome}{grouping variable}

\item{levels}{levels of the grouping variable that are to be classified (should be of length 2)}

\item{cutoffs.1}{cutoff values to be shown in the ROC curve}

\item{cutoffs.2}{cutoff values to be shown in the sensitivity/specificity plot}

\item{cutoffs.3}{cutoff values to be shown in the conditional distribution plot}

\item{xlab.3}{label for x axis in conditional distribution plot}

\item{labels.3}{legend labels for both groups in \code{levels} in conditional distribution plot}

\item{xlim.3}{\code{xlim} for conditional distribution plot}

\item{ylim.3}{\code{ylim} for conditional distribution plot}
}
\value{
\code{classificationplots} returns three plots, and invisibly a dataframe containing sensitivities and specificities at all available threshold values.
}
\description{
Generate plots for classification of two groups given one continuous variable
}
\examples{
library(pROC)
classificationplots(aSAH$s100b,aSAH$outcome,levels=c("Good","Poor"))
}
\author{
Mathijs Deen
}
